requirements.txt
Django==2.2.8
django-environ==0.4.5
django-allauth==0.39.1
django-filter==2.1.0
django-uuslug==1.2.0
djangorestframework==3.9.4
Markdown==3.1.1
Pillow==6.2.1
six==1.12.0
django-oss-storage==1.1.0